// types/rules-setting.interface.ts
export type RulesSettingType = {
    min: string;
    max?: string;
    uppercase: boolean;
    lowercase: boolean;
    numbers: boolean;
    symbols: boolean;
    favorite_food?: string | undefined;
    favorite_drink?: string | undefined;
    favorite_place?: string | undefined;
    favorite_person?: string | undefined;
};